main만 참고 나머지 확인 못함
