+++ Test10/.idea/misc.xml
@@ -3,6 +3,8 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/jdh33/MyGitRepo/foodmap/Test10/app/src/main/res/layout/activity_login.xml" value="0.11956521739130435" />
+        <entry key="..\:/Users/jdh33/MyGitRepo/foodmap/Test10/app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="..\:/Users/user/MygitRepo/foodmap/Test10/app/src/main/res/layout/activity_join.xml" value="0.20364583333333333" />
         <entry key="..\:/Users/user/MygitRepo/foodmap/Test10/app/src/main/res/layout/activity_login.xml" value="0.20364583333333333" />
       </map>
